move config code out of modules directory
